Fix warning on WIN

+++ test/alt-util_test.cpp
@@ -10,7 +10,7 @@ struct AltUtilTest : public ::testing::Test {
 };
 
 TEST_F(AltUtilTest, GetLastKnownBlocks) {
-  auto genesis = getLastKnownBlocks(popminer.btc(), 1);
+  auto genesis = getLastKnownBlocks(popminer.btc(), 1ull);
   ASSERT_EQ(genesis.size(), 1);
   ASSERT_EQ(genesis[0],
             popminer.getBtcParams().getGenesisBlock().getHash().asVector());
@@ -20,12 +20,12 @@ TEST_F(AltUtilTest, GetLastKnownBlocks) {
   ASSERT_EQ(one.size(), 1);
   ASSERT_EQ(one[0], a->getHash().asVector());
 
-  auto ten = getLastKnownBlocks(popminer.btc(), 10);
+  auto ten = getLastKnownBlocks(popminer.btc(), 10ull);
   ASSERT_EQ(ten.size(), 10);
   for (size_t i = 0; i < 10; i++) {
-    ASSERT_EQ(ten[i], a->getAncestor(1 + i)->getHash().asVector());
+    ASSERT_EQ(ten[i], a->getAncestor((int)(1 + i))->getHash().asVector());
   }
 
-  auto many = getLastKnownBlocks(popminer.btc(), 1000);
-  ASSERT_EQ(many.size(), 11);  // we know only 11 bocks, 10+genesis
+  auto many = getLastKnownBlocks(popminer.btc(), 1000ull);
+  ASSERT_EQ(many.size(), 11ull);  // we know only 11 bocks, 10+genesis
 }
